integrate XCTest with Spectre for nicer tests in Xcode
