+++ Package.resolved
@@ -42,8 +42,8 @@
         "repositoryURL": "https://github.com/onevcat/Rainbow.git",
         "state": {
           "branch": null,
-          "revision": "f69961599ad524251d677fbec9e4bac57385d6fc",
-          "version": "3.1.1"
+          "revision": "797a68d0a642609424b08f11eb56974a54d5f6e2",
+          "version": "3.1.4"
         }
       },
       {
@@ -57,11 +57,11 @@
       },
       {
         "package": "Spectre",
-        "repositoryURL": "https://github.com/kylef/Spectre.git",
+        "repositoryURL": "https://github.com/yonaskolb/Spectre.git",
         "state": {
           "branch": null,
-          "revision": "e34d5687e1e9d865e3527dd58bc2f7464ef6d936",
-          "version": "0.8.0"
+          "revision": "2b520ec417fc515c70e7093ae3ee321c3237e215",
+          "version": "0.8.1"
         }
       },
       {
