+++ Package.swift
@@ -14,7 +14,7 @@ let package = Package(
         .package(url: "https://github.com/kylef/Commander.git", from: "0.8.0"),
         .package(url: "https://github.com/jpsim/Yams.git", from: "0.3.6"),
         .package(url: "https://github.com/yonaskolb/JSONUtilities.git", from: "3.3.0"),
-        .package(url: "https://github.com/kylef/Spectre.git", from: "0.8.0"),
+        .package(url: "https://github.com/yonaskolb/Spectre.git", from: "0.8.1"),
         .package(url: "https://github.com/onevcat/Rainbow.git", from: "3.0.0"),
         .package(url: "https://github.com/xcodeswift/xcproj.git", from: "4.1.0")
     ],
