--- Tests/XcodeGenKitTests/XCTest.swift
@@ -1,12 +1,0 @@
-import Foundation
-import XCTest
-
-class XCodeGenKitTests: XCTestCase {
-
-    func testXcodeGenKit() {
-        projectGeneratorTests()
-        projectLoadingTests()
-        fixtureTests()
-        projectSpecTests()
-    }
-}
